python/pyproject.toml
[project]
name = "alaska-legislative-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fire>=0.7.0",
    "httpx>=0.28.1",
    "ibis-framework[duckdb]>=10.0.0",
    "ipykernel>=6.29.5",
    "pytest>=8.3.4",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
